requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob


pdf2image
# !apt-get install poppler-utils 
# !sudo apt install tesseract-ocr
pytesseract
reportlab
Pillow==9.0.0
PyPDF2
opencv-python
pandas
python-dateutil